LayoutTests/platform/ios-sim/fast/multicol/hit-test-above-or-below-expected.txt

CONSOLE MESSAGE: line 9: TypeError: 'null' is not an object (evaluating 'range.startContainer')
123
abc
def
ghi
jkl
mno
